ref: externalized `EQResulttTable` drawer
